Safari-609.1.11.1/LayoutTests/fast/xpath/invalid-functions-expected.txt

PASS document.createExpression("foobar()", null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.evaluate("foobar()", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.createExpression("position(//div)", null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.evaluate("position(//div)", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.createExpression("ceiling(1,2)", null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.evaluate("ceiling(1,2)", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.createExpression("ceiling(1,2)", null).evaluate(document, XPathResult.ANY_TYPE, null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.createExpression("boolean()", null) threw exception SyntaxError: The string did not match the expected pattern..
PASS document.evaluate("boolean()", document, null, XPathResult.ANY_TYPE, null) threw exception SyntaxError: The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
